+++ external/game/model.go
@@ -1,0 +1,20 @@
+package game
+
+import (
+	"time"
+
+	"github.com/doruo/falloutdle/internal/character"
+)
+
+// Game represents a current game state
+type Game struct {
+	CurrentCharacter character.Character
+	Date             time.Time `json:"date"`
+}
+
+func NewGame(c character.Character) *Game {
+	return &Game{
+		CurrentCharacter: c,
+		Date:             time.Now(),
+	}
+}
